Add internal customer ID (#26)
As we transition away from using customer slugs as part of CUR ingestion bucket names, this diff introduces the concept of an internal customer ID (based on the customer's randomly-generated external ID). This internal ID will eventually replace the customer slug entirely in these scripts. To err on the side of caution, I've granted access to both the old and new bucket names for now.